+++ modules/multiplayer/editor/multiplayer_editor_plugin.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  replication_editor_plugin.h                                          */
+/*  multiplayer_editor_plugin.h                                          */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,89 +28,39 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef REPLICATION_EDITOR_PLUGIN_H
-#define REPLICATION_EDITOR_PLUGIN_H
+#ifndef MULTIPLAYER_EDITOR_PLUGIN_H
+#define MULTIPLAYER_EDITOR_PLUGIN_H
 
 #include "editor/editor_plugin.h"
-#include "modules/multiplayer/scene_replication_config.h"
-#include "scene/gui/box_container.h"
 
-class ConfirmationDialog;
-class MultiplayerSynchronizer;
-class AcceptDialog;
-class LineEdit;
-class Tree;
-class TreeItem;
-class PropertySelector;
-class SceneTreeDialog;
+#include "editor/plugins/editor_debugger_plugin.h"
 
-class ReplicationEditor : public VBoxContainer {
-	GDCLASS(ReplicationEditor, VBoxContainer);
+class EditorNetworkProfiler;
+class MultiplayerEditorDebugger : public EditorDebuggerPlugin {
+	GDCLASS(MultiplayerEditorDebugger, EditorDebuggerPlugin);
 
 private:
-	MultiplayerSynchronizer *current = nullptr;
+	HashMap<int, EditorNetworkProfiler *> profilers;
 
-	AcceptDialog *error_dialog = nullptr;
-	ConfirmationDialog *delete_dialog = nullptr;
-	Button *add_pick_button = nullptr;
-	Button *add_from_path_button = nullptr;
-	LineEdit *np_line_edit = nullptr;
-
-	Label *drop_label = nullptr;
-
-	Ref<SceneReplicationConfig> config;
-	NodePath deleting;
-	Tree *tree = nullptr;
-
-	PropertySelector *prop_selector = nullptr;
-	SceneTreeDialog *pick_node = nullptr;
-	NodePath adding_node_path;
-
-	Button *pin = nullptr;
-
-	Ref<Texture2D> _get_class_icon(const Node *p_node);
-
-	void _add_pressed();
-	void _tree_item_edited();
-	void _tree_button_pressed(Object *p_item, int p_column, int p_id, MouseButton p_button);
-	void _update_checked(const NodePath &p_prop, int p_column, bool p_checked);
-	void _update_config();
-	void _dialog_closed(bool p_confirmed);
-	void _add_property(const NodePath &p_property, bool p_spawn = true, bool p_sync = true);
-
-	void _pick_node_filter_text_changed(const String &p_newtext);
-	void _pick_node_select_recursive(TreeItem *p_item, const String &p_filter, Vector<Node *> &p_select_candidates);
-	void _pick_node_filter_input(const Ref<InputEvent> &p_ie);
-	void _pick_node_selected(NodePath p_path);
-
-	void _pick_new_property();
-	void _pick_node_property_selected(String p_name);
-
-	bool _can_drop_data_fw(const Point2 &p_point, const Variant &p_data, Control *p_from) const;
-	void _drop_data_fw(const Point2 &p_point, const Variant &p_data, Control *p_from);
-
-	void _add_sync_property(String p_path);
-
-protected:
-	static void _bind_methods();
-
-	void _notification(int p_what);
+	void _profiler_activate(bool p_enable, int p_session_id);
 
 public:
-	void edit(MultiplayerSynchronizer *p_object);
-	MultiplayerSynchronizer *get_current() const { return current; }
+	virtual bool has_capture(const String &p_capture) const override;
+	virtual bool capture(const String &p_message, const Array &p_data, int p_index) override;
+	virtual void setup_session(int p_session_id) override;
 
-	Button *get_pin() { return pin; }
-	ReplicationEditor();
-	~ReplicationEditor() {}
+	MultiplayerEditorDebugger() {}
 };
 
-class ReplicationEditorPlugin : public EditorPlugin {
-	GDCLASS(ReplicationEditorPlugin, EditorPlugin);
+class ReplicationEditor;
+
+class MultiplayerEditorPlugin : public EditorPlugin {
+	GDCLASS(MultiplayerEditorPlugin, EditorPlugin);
 
 private:
 	Button *button = nullptr;
 	ReplicationEditor *repl_editor = nullptr;
+	Ref<MultiplayerEditorDebugger> debugger;
 
 	void _node_removed(Node *p_node);
 
@@ -124,8 +74,8 @@ public:
 	virtual bool handles(Object *p_object) const override;
 	virtual void make_visible(bool p_visible) override;
 
-	ReplicationEditorPlugin();
-	~ReplicationEditorPlugin();
+	MultiplayerEditorPlugin();
+	~MultiplayerEditorPlugin();
 };
 
-#endif // REPLICATION_EDITOR_PLUGIN_H
+#endif // MULTIPLAYER_EDITOR_PLUGIN_H
